refactor: delete useless data folder in mainpage
