+++ lib/app/pages/mainpage/ui/navermap/route_jongro07.dart
@@ -1,26 +1,26 @@
 import 'package:flutter_naver_map/flutter_naver_map.dart';
 
-const stations = [
-  NLatLng(37.587707, 126.996686),
-  NLatLng(37.589185, 126.9972903),
-  NLatLng(37.589829, 126.998893),
-  NLatLng(37.587254, 126.999908),
-  NLatLng(37.586098, 127.000658),
-  NLatLng(37.58350804, 127.0015539),
-  NLatLng(37.581117, 127.001914),
-  NLatLng(37.578845, 127.002061),
-  NLatLng(37.576628, 127.002098),
-  NLatLng(37.578149, 127.002411),
-  NLatLng(37.58111949, 127.0022769),
-  NLatLng(37.58348551, 127.0018454),
-  NLatLng(37.58558033, 127.0006878),
-  NLatLng(37.583323, 126.998977),
-  NLatLng(37.584897, 126.996569),
-  NLatLng(37.5873281, 126.9939426),
-  NLatLng(37.587943, 126.992523),
-  NLatLng(37.58744522, 126.9937417),
-  NLatLng(37.58516359, 126.9971922),
-];
+// const stations = [
+//   NLatLng(37.587707, 126.996686),
+//   NLatLng(37.589185, 126.9972903),
+//   NLatLng(37.589829, 126.998893),
+//   NLatLng(37.587254, 126.999908),
+//   NLatLng(37.586098, 127.000658),
+//   NLatLng(37.58350804, 127.0015539),
+//   NLatLng(37.581117, 127.001914),
+//   NLatLng(37.578845, 127.002061),
+//   NLatLng(37.576628, 127.002098),
+//   NLatLng(37.578149, 127.002411),
+//   NLatLng(37.58111949, 127.0022769),
+//   NLatLng(37.58348551, 127.0018454),
+//   NLatLng(37.58558033, 127.0006878),
+//   NLatLng(37.583323, 126.998977),
+//   NLatLng(37.584897, 126.996569),
+//   NLatLng(37.5873281, 126.9939426),
+//   NLatLng(37.587943, 126.992523),
+//   NLatLng(37.58744522, 126.9937417),
+//   NLatLng(37.58516359, 126.9971922),
+// ];
 
 const jongroRoute = [
   NLatLng(37.587707, 126.996686), // 명륜새마을금고
